Add files via upload
School culture construction is a significant topic in China's educational modernization and school development, as well as an important component of teacher-school educational practices. This study aims to explore the role and function of teachers in school culture construction, with a focus on analyzing the process of teacher role construction from the perspective of teacher empowerment. It is essential to clarify the educational practices of teacher-school culture construction and promote teachers' professional autonomy. J County Schools, as newly established educational experimental schools in China, are collaborating with University B and J County Education Research Institute to advance the theoretical conceptualization and practical construction of J School culture, with teachers as the main agents for capacity building and practical activities. Using role theory and teacher empowerment theory as the theoretical framework, this study conducts a case study based on teachers from J County Group Schools. Data on the roles and empowerment of teachers in school culture construction are collected and analyzed through interviews, physical collection methods, and semi-participatory observation. The study identifies six typical roles of teachers in school culture, namely "practitioner and innovator," "organizer and facilitator," "learner and miner," "reflector and maintainer," "interlocutor and sharer," and "collaborator and communicator." The corresponding functions of these roles are summarized. The study also analyzes the process of teacher role construction through the perception of psychological and structural empowerment practices, identifying the empowering effects of "guidance, adjustment, and reconciliation" in promoting teacher role practices. Furthermore, it proposes empowerment strategies focused on improving teachers' perception of structural and psychological empowerment in relation to school leadership. Taking the school culture construction practices of newly established schools in Chinese counties as a typical example, this study outlines the development path of school culture construction in county schools under the guidance of collaboration between universities and local education research institutes. It establishes six major roles and their functions for teachers participating in school culture construction, and analyzes the process of teacher role construction from an empowerment perspective. The findings aim to provide insights and guidance for teacher participation in school culture construction, particularly in terms of role functions and empowerment strategies.